run both scheduler and worker for jobs
This changes the way the run method works on jobs: now both the
scheduler and the worker will be spawned off into separate tasks and
selected over.

Additionally a new run_worker method is provided as a convenience, which
is the complement to run_scheduler.

By doing so, in-process task execution is simplified when using
scheduled jobs.