Step functions (#24)
This introduces step functions.

Step functions are a series of linked tasks where each task takes input from the output of previous one. Said another way, step functions are a chain of dependent tasks, each depending on its immediate ancestor.

This design makes it possible to define a series of discrete tasks that can perform work which is required for further work as the series progresses. What this means is that when tasks fail and are retried, prior work is not repeated.

With this patch jobs are now composed of one or more step functions.

Enabling this interface also necessitates updating tasks to:

Provide an output as an associated type and,
Take a transaction as input to their execute method.
It's worth noting that various breaking changes are contained throughout and this is a significant change to the job API especially.

Note that on series 0.0.x we are NOT adhering to proper semver. Please upgrade with caution until the 0.1.x series.