simplify scheduler somewhat
Here we iterate the retrieved schedule directly, waiting until the next
enqueue or we receive a shutdown signal, whichever happens first.

If the schedule ends, then the scheduler will return and no more
processing will occur until a new schedule is set and a new scheduler is
running.

In the future we may want to support updating schedules with a trigger a
la the worker.