provide task_attempt table (#50)
This reworks how task executions are tracked by providing a new table,
`task_attempt` which is a log of each dequeue of a given task from the
queue. Put another way, each time a task is executed, a row is inserted
into the attempts table and the result of that execution is stored on
that row.

As such, the `task` table is now a container for the fixed metadata of a
specific enqueue of a task. This means that metadata associated with an
attempt now lives on the `task_attempt` table. For example, error
messages are stored on the attempts table and likewise creation and
completion timestamps also live on this table.

Importantly some fields of the `task` row have been removed as they are
now derived from the attempts log. For instance, `task_attempt` provides
an `updated_at` column which allows for e.g. determining `succeeded_at`
or `failed_at` timestamps of the task enqueue. Similarly `retry_count`
is a function of the number of attempt rows associated to a given task.

Furthermore, the retry policy is now embedded as a JSONB column on the
task row.

Related discussion: https://github.com/maxcountryman/underway/discussions/17