ensure task attempts metadata is correct (#54)
This fixes an issue where every attempt row would be updated when
reaching a terminal state. Here we address this by only updating the
most recent, i.e. current, attempt.

We also address an issue where errors occurring in a job would not
persist pending attempts because the transaction would not be
committed.

Regarding this last fix, it's possible that the transaction commit
should be relocated to the worker such that no commit is required in the
job--this should be addressed separately.