stale dequeues consider retries
This fixes an issue where dequeuing a stall retry would not consider if
the related task had available retries or not.

To address this we rework the dequeue query to check the retry record
directly as part of the condition for selecting a stale task.

We also provide an update over attempt rows when a task row is selected
to ensure that prior in-progress attempts are marked as failed.