refactor scheduler iterator
This refactors the scheduler iterator into its own type which manages
the upcoming schedule directly. At the same time, this iterator now
produces `Zoned` values instead of calculating relative durations.

It also provides a function, `wait_until`, which consumes `Zoned` and
checks if the current time indicates we're done waiting relative to the
provided time.

When the provided time is in the future, we wait until that instant and
then make the same check again.