provide test for heartbeat (#65)
This is a rough test of the way the heartbeat process operates. It
checks that the heartbeat is updating the associated task row as we'd
expect as well as stops updating the same row after execution.

Unfortunately it's somewhat difficult to test this behavior and arguably
this isn't the correct place to do so. Even so, this is crucial to the
correctness of the system and so something here is better than nothing.