From<Job> for Worker and Scheduler
Since `Job` has all its attributes private to the module, the only way to create a worker or
scheduler is to either keep a reference to the queue used in the job creation.

Implementing `From<Job<I, S>>` for `Worker` and for `Scheduler` makes it so all we need is the
job reference.